requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
sentry-sdk[fastapi]==1.39.2
slowapi==0.1.9
pytest==8.3.3
